+++ inlang/source-code/settings-component/src/stories/inlang-settings.ts
@@ -38,8 +38,6 @@ export default class InlangSettings extends LitElement {
 		css`
 			h3 {
 				margin: 0;
-				padding-bottom: 24px;
-				border-bottom: 1px solid var(--sl-panel-border-color);
 			}
 			.container {
 				position: relative;
@@ -50,7 +48,7 @@ export default class InlangSettings extends LitElement {
 			.module-container {
 				display: flex;
 				flex-direction: column;
-				gap: 24px;
+				gap: 16px;
 			}
 			.hover-bar-container {
 				width: 100%;
@@ -82,8 +80,26 @@ export default class InlangSettings extends LitElement {
 		`,
 	]
 
-	@property({ attribute: false })
-	project: InlangProject = {} as InlangProject
+	@property({ type: Object })
+	settings: ReturnType<InlangProject["settings"]> = {} as ReturnType<InlangProject["settings"]>
+
+	@property({ type: Array })
+	installedPlugins: ReturnType<InlangProject["installed"]["plugins"]> = [] as ReturnType<
+		InlangProject["installed"]["plugins"]
+	>
+
+	@property({ type: Array })
+	installedMessageLintRules: ReturnType<InlangProject["installed"]["messageLintRules"]> =
+		[] as ReturnType<InlangProject["installed"]["messageLintRules"]>
+
+	dispatchOnSetSettings(settings: ProjectSettings) {
+		const onSetSettings = new CustomEvent("setSettings", {
+			detail: {
+				argument: settings,
+			},
+		})
+		this.dispatchEvent(onSetSettings)
+	}
 
 	@state()
 	private _newSettings: ProjectSettings | undefined = undefined
@@ -94,8 +110,8 @@ export default class InlangSettings extends LitElement {
 	override async firstUpdated() {
 		await this.updateComplete
 
-		if (this.project?.settings()) {
-			this._newSettings = JSON.parse(JSON.stringify(this.project.settings()))
+		if (this.settings) {
+			this._newSettings = JSON.parse(JSON.stringify(this.settings))
 		}
 	}
 
@@ -120,7 +136,7 @@ export default class InlangSettings extends LitElement {
 				[property]: value,
 			}
 		}
-		if (JSON.stringify(this.project?.settings()) !== JSON.stringify(this._newSettings)) {
+		if (JSON.stringify(this.settings) !== JSON.stringify(this._newSettings)) {
 			this._unsavedChanges = true
 		} else {
 			this._unsavedChanges = false
@@ -128,29 +144,32 @@ export default class InlangSettings extends LitElement {
 	}
 
 	_revertChanges = () => {
-		if (this.project?.settings()) {
-			this._newSettings = JSON.parse(JSON.stringify(this.project?.settings()))
+		if (this.settings) {
+			this._newSettings = JSON.parse(JSON.stringify(this.settings))
 		}
 		this._unsavedChanges = false
 	}
 
 	_saveChanges = () => {
 		if (this._newSettings) {
-			this.project?.setSettings(this._newSettings)
+			this.dispatchOnSetSettings(this._newSettings)
 		}
 		this._unsavedChanges = false
 	}
 
-	private get _projectSettingProperties(): Record<
+	private get _settingProperties(): Record<
 		InlangModule["default"]["id"] | "internal",
 		{
 			meta?: InstalledPlugin | InstalledMessageLintRule
 			schema?: Record<string, Record<string, unknown>>
 		}
 	> {
-		const _project = this.project
-		if (!_project) throw new Error("No inlang project")
-		if (!_project.settings()) throw new Error("No inlang project settings")
+		const _settings = this.settings
+		const _installedPlugins = this.installedPlugins
+		const _installedMessageLintRules = this.installedMessageLintRules
+
+		if (!_settings) throw new Error("No inlang settings")
+		if (!_installedPlugins) throw new Error("No installed plugins")
 
 		const generalSchema: Record<
 			InlangModule["default"]["id"] | "internal",
@@ -160,7 +179,7 @@ export default class InlangSettings extends LitElement {
 			}
 		> = { internal: { schema: ProjectSettings.allOf[0] } }
 
-		for (const plugin of _project.installed.plugins()) {
+		for (const plugin of _installedPlugins) {
 			if (plugin.settingsSchema) {
 				generalSchema[plugin.id] = {
 					schema: plugin.settingsSchema,
@@ -168,7 +187,7 @@ export default class InlangSettings extends LitElement {
 				}
 			}
 		}
-		for (const lintRule of _project.installed.messageLintRules()) {
+		for (const lintRule of _installedMessageLintRules) {
 			if (lintRule.settingsSchema) {
 				generalSchema[lintRule.id] = {
 					schema: lintRule.settingsSchema,
@@ -182,10 +201,12 @@ export default class InlangSettings extends LitElement {
 
 	override render() {
 		return html` <div class="container">
-			${Object.entries(this._projectSettingProperties).map(([key, value]) => {
+			${Object.entries(this._settingProperties).map(([key, value]) => {
 				return value.schema?.properties && this._newSettings
 					? html`<div class="module-container">
-							<h3>${(value.meta && (value.meta?.displayName as { en: string }).en) || key}</h3>
+							${value.meta &&
+							(value.meta?.displayName as { en: string }).en &&
+							html`<h3>${value.meta && (value.meta?.displayName as { en: string }).en}</h3>`}
 							${Object.entries(value.schema.properties).map(([property, schema]) => {
 								if (property === "$schema" || property === "modules" || property === "experimental")
 									return undefined
@@ -193,7 +214,7 @@ export default class InlangSettings extends LitElement {
 									? html`
 											<general-input
 												.property=${property}
-												.modules=${this.project.installed.messageLintRules() || []}
+												.modules=${this.installedMessageLintRules || []}
 												.value=${this._newSettings?.[property as keyof typeof this._newSettings]}
 												.schema=${schema}
 												.handleInlangProjectChange=${this.handleInlangProjectChange}