+++ inlang/source-code/sdk/load-test/CHANGELOG.md
@@ -4,6 +4,32 @@
 
 ### Patch Changes
 
+- Updated dependencies [d1b361e]
+  - @lix-js/client@1.1.0
+  - @inlang/cli@2.13.3
+  - @inlang/sdk@0.28.2
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [03fa6f2]
+  - @lix-js/client@1.0.0
+  - @inlang/cli@2.13.2
+  - @inlang/sdk@0.28.1
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [1e43ae4]
+  - @inlang/sdk@0.28.0
+  - @inlang/cli@2.13.1
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [4837297]
   - @inlang/cli@2.13.0
   - @inlang/sdk@0.27.0